refactor(project-api): split client.ts into command-family modules
Move the 1155-line api/http/project/client.ts into api/http/project/client/
with one module per existing section banner (orgs, collabSync, projects,
labels, milestones, members, workItems, workRuns, discussions,
workItemProperties, routineWebhooks, batch, routines, assets) plus a shared
scope.ts. client/index.ts re-exports the same surface so "./client" still
resolves; client.test.ts moves to client/workItems.test.ts. Pure code motion.
